move game states to new file
